requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
pydantic
python-dotenv
openpyxl
gspread
google-auth
passlib
python-jose[cryptography]